Asset Entry Entry and List Working Fine

+++ AssetDetail/src/main/java/com/asset/asset/AssetRepository.java
@@ -1,9 +1,10 @@
 package com.asset.asset;
 
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AssetRepository extends JpaRepository<Asset, Long> {
+public interface AssetRepository extends JpaRepository<Asset, Integer> {
 
 }
